+++ src/seamlesserverpkg/renderer/base/header.nim
@@ -1,6 +1,16 @@
 include pkg/karax/prelude
 
-proc renderHeader*: VNode =
+import ./state
+import ./karaxNodes/dynamicLink
+
+proc renderHeader*(state: State): VNode =
   ## Render page header
   buildHtml(header):
-    text "header"
+    h2: text state.brData.appName
+    ul(class = "nav"):
+      li: dynamicLink(href = "/"):
+        text "Home"
+      li: dynamicLink(href = "/user/login"):
+        text "Login"
+      li: dynamicLink(href = "/user/register"):
+        text "Register"