+++ src/seamlesserverpkg/renderer/routes/user/login.nim
@@ -10,8 +10,20 @@ proc renderHtml*(state: State): Rendered =
   result.title = "Login"
   result.vnode = buildHtml(main):
     h1: text "Login page"
-    dynamicLink(href = "/"):
-      text "Homepage"
+    form(`method` = "post"):
+      input(
+        name = "username",
+        placeholder = "Username",
+        `type` = "text"
+      )
+      input(
+        name = "password",
+        placeholder = "Password",
+        `type` = "password"
+      )
+      button(`type` = "submit"): text "Submit"
+
+    
 
 when not defined js:
   import pkg/prologue