+++ changelog.md
@@ -5,6 +5,9 @@
 - Removed debug middleware when not in debug mode
 - Added configurable JS filename
 - Improved and added more nimble tasks
+- Header and footer receives the state
+- Added links to header
+- Added login form HTML
 
 ## Version alpha 0.0.2 (27/06/2023)